+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.0.0)
 
 
+### 🔔 变化
+
+- **buession-security-web：** 安全配置类中所有基本类型，调整为包装类型 
+
+
 ### 🐞 Bug 修复
 
 - **buession-security-web：** 修复 ReactiveWebSecurityConfigurerAdapterConfiguration 无参数构造函数，为初始化 Configurer 的 BUG